requirements.txt
click==8.1.7
rich==13.7.1
httpx==0.27.0
tenacity==8.2.3
pydantic==2.6.4
python-dotenv==1.0.1
reportlab==4.1.0
jinja2==3.1.3
openai==1.30.1
pandas==2.2.2
openpyxl==3.1.2
loguru==0.7.2
tabulate==0.9.0
colorama==0.4.6
pytest==8.2.0